Move type definitions into separate files
