+++ source/type/IdeaSeverity.ts
@@ -1,0 +1,9 @@
+// https://hackage.haskell.org/package/hlint-3.6.1/docs/Language-Haskell-HLint.html#t:Severity
+enum IdeaSeverity {
+  Ignore = "Ignore",
+  Suggestion = "Suggestion",
+  Warning = "Warning",
+  Error = "Error",
+}
+
+export default IdeaSeverity;
