+++ source/type/Key.ts
@@ -1,0 +1,3 @@
+type Key = string;
+
+export default Key;
