+++ source/type/MessageReason.ts
@@ -1,0 +1,4 @@
+// https://downloads.haskell.org/ghc/9.6.2/docs/libraries/ghc-9.6.2/GHC-Driver-Flags.html#t:WarningFlag
+type MessageReason = string;
+
+export default MessageReason;
